refactor(ui): make create alert single page more generic to use with edit
