+++ thirdeye-ui/src/app/components/alert-wizard-v2/alert-notifications/alert-notifications.interfaces.ts
@@ -17,4 +17,5 @@ import { SubscriptionGroup } from "../../../rest/dto/subscription-group.interfac
 export interface AlertNotificationsProps {
     alert: EditableAlert;
     onSubscriptionGroupsChange: (newGroups: SubscriptionGroup[]) => void;
+    initiallySelectedSubscriptionGroups: SubscriptionGroup[];
 }
