+++ thirdeye-ui/src/app/pages/alerts-update-page/alerts-update-page.interfaces.ts
@@ -1,3 +1,6 @@
+import { EditableAlert } from "../../rest/dto/alert.interfaces";
+import { SubscriptionGroup } from "../../rest/dto/subscription-group.interfaces";
+
 /**
  * Copyright 2022 StarTree Inc
  *
@@ -14,3 +17,12 @@
 export type AlertsUpdatePageParams = {
     id: string;
 };
+
+export interface AlertsEditPageProps {
+    submitButtonLabel: string;
+    startingAlertConfiguration: EditableAlert;
+    pageTitle: string;
+    onSubmit: (alert: EditableAlert) => void;
+    selectedSubscriptionGroups: SubscriptionGroup[];
+    onSubscriptionGroupChange: (newGroups: SubscriptionGroup[]) => void;
+}
