+++ thirdeye-ui/src/app/pages/alerts-update-page/alerts-update-simple-page.component.tsx
@@ -14,18 +14,18 @@
 import { Grid } from "@material-ui/core";
 import React, { FunctionComponent } from "react";
 import { useOutletContext } from "react-router-dom";
-import { AlertDetails } from "../../../components/alert-wizard-v2/alert-details/alert-details.component";
-import { AlertNotifications } from "../../../components/alert-wizard-v2/alert-notifications/alert-notifications.component";
-import { AlertTemplate } from "../../../components/alert-wizard-v2/alert-template/alert-template.component";
-import { AlertTemplate as AlertTemplateType } from "../../../rest/dto/alert-template.interfaces";
-import { EditableAlert } from "../../../rest/dto/alert.interfaces";
-import { SubscriptionGroup } from "../../../rest/dto/subscription-group.interfaces";
+import { AlertDetails } from "../../components/alert-wizard-v2/alert-details/alert-details.component";
+import { AlertNotifications } from "../../components/alert-wizard-v2/alert-notifications/alert-notifications.component";
+import { AlertTemplate } from "../../components/alert-wizard-v2/alert-template/alert-template.component";
+import { AlertTemplate as AlertTemplateType } from "../../rest/dto/alert-template.interfaces";
+import { EditableAlert } from "../../rest/dto/alert.interfaces";
+import { SubscriptionGroup } from "../../rest/dto/subscription-group.interfaces";
 
-export const AlertsCreateSimplePage: FunctionComponent = () => {
+export const AlertsUpdateSimplePage: FunctionComponent = () => {
     const [
         alert,
         onAlertPropertyChange,
-        ,
+        selectedSubscriptionGroups,
         onSubscriptionGroupsChange,
         selectedAlertTemplate,
         setSelectedAlertTemplate,
@@ -63,7 +63,9 @@ export const AlertsCreateSimplePage: FunctionComponent = () => {
             <Grid item xs={12}>
                 <AlertNotifications
                     alert={alert}
-                    initiallySelectedSubscriptionGroups={[]}
+                    initiallySelectedSubscriptionGroups={
+                        selectedSubscriptionGroups
+                    }
                     onSubscriptionGroupsChange={onSubscriptionGroupsChange}
                 />
             </Grid>
