+++ thirdeye-ui/src/app/utils/routes/routes.util.ts
@@ -32,7 +32,9 @@ export const AppRouteRelative = {
     ALERTS_CREATE_COPY: `copy/${PLACEHOLDER_ROUTE_ID}`,
     ALERTS_CREATE_SIMPLE: "simple",
     ALERTS_CREATE_ADVANCED: "advanced",
-    ALERTS_UPDATE: `update`,
+    ALERTS_UPDATE: "update",
+    ALERTS_UPDATE_SIMPLE: "simple",
+    ALERTS_UPDATE_ADVANCED: "advanced",
     ANOMALIES: "anomalies",
     ANOMALIES_ALL: "all",
     ANOMALIES_ALL_RANGE: "range",
@@ -91,6 +93,8 @@ export const AppRoute = {
     ALERTS_CREATE_EXISTING_ADVANCED: `/${AppRouteRelative.ALERTS}/${AppRouteRelative.ALERTS_CREATE}/${AppRouteRelative.ALERTS_CREATE_COPY}/${AppRouteRelative.ALERTS_CREATE_ADVANCED}`,
     ALERTS_ALERT: `/${AppRouteRelative.ALERTS}/${AppRouteRelative.ALERTS_ALERT}`,
     ALERTS_UPDATE: `/${AppRouteRelative.ALERTS}/${AppRouteRelative.ALERTS_ALERT}/${AppRouteRelative.ALERTS_UPDATE}`,
+    ALERTS_UPDATE_SIMPLE: `/${AppRouteRelative.ALERTS}/${AppRouteRelative.ALERTS_ALERT}/${AppRouteRelative.ALERTS_UPDATE}/${AppRouteRelative.ALERTS_CREATE_SIMPLE}`,
+    ALERTS_UPDATE_ADVANCED: `/${AppRouteRelative.ALERTS}/${AppRouteRelative.ALERTS_ALERT}/${AppRouteRelative.ALERTS_UPDATE}/${AppRouteRelative.ALERTS_CREATE_ADVANCED}`,
     ANOMALIES: `/${AppRouteRelative.ANOMALIES}`,
     ANOMALIES_ALL: `/${AppRouteRelative.ANOMALIES}/${AppRouteRelative.ANOMALIES_ALL}`,
     ANOMALIES_ALL_RANGE: `/${AppRouteRelative.ANOMALIES}/${AppRouteRelative.ANOMALIES_ALL}/${AppRouteRelative.ANOMALIES_ALL_RANGE}`,
